requirements.txt
alembic==1.3.2
aniso8601==8.0.0
asgiref==3.2.3
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
Django==3.0.2
earthengine-api==0.1.210
Faker==3.0.0
Fiona==1.8.13
Flask==1.1.1
Flask-Admin==1.5.4
Flask-Caching==1.8.0
Flask-Flash==1.6.8
Flask-HTTPAuth==3.3.0
Flask-Login==0.4.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
future==0.18.2
geopandas==0.6.2
google-api-core==1.15.0
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
google-cloud-bigquery==1.23.1
google-cloud-core==1.1.0
google-cloud-storage==1.24.1
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
gunicorn==20.0.4
httplib2==0.15.0
idna==2.8
image==1.5.27
importlib-metadata==1.3.0
inflect==4.0.0
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
matplotlib==3.1.2
more-itertools==8.0.2
munch==2.5.0
numpy==1.18.1
pandas==0.25.3
Pillow==7.0.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
pyparsing==2.4.6
pyproj==2.4.2.post1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3
requests==2.22.0
rsa==4.0
Shapely==1.6.4.post2
six==1.13.0
SQLAlchemy==1.3.12
sqlparse==0.3.0
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.7
Werkzeug==0.16.0
WTForms==2.2.1
zipp==0.6.0
